update to headless server base 2404
