--- headless-server-commerce-app/src/main/resources/logback-common.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<included>
-  <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
-  <!--Workaround to get the property injected. See https://github.com/spring-cloud/spring-cloud-sleuth/issues/330-->
-  <springProperty scope="context" name="springAppName" source="spring.application.name"/>
-  <property name="LOG_FILE" value="${LOG_FILE:-${LOG_PATH:-${LOG_TEMP:-${java.io.tmpdir:-/tmp}}}/${springAppName:-spring}.log}"/>
-  <include resource="org/springframework/boot/logging/logback/console-appender.xml"/>
-  <include resource="org/springframework/boot/logging/logback/file-appender.xml"/>
-
-</included>
