+++ headless-server-commerce-base-lib/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 com.coremedia.blueprint.headlessserver.CommerceCaasAutoConfiguration
+com.coremedia.blueprint.headlessserver.CorsAutoConfiguration
