+++ headless-server-commerce-lib/pom.xml
@@ -1,19 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-
   <parent>
     <groupId>com.coremedia.commerce.labs</groupId>
     <artifactId>headless-server-commerce-parent</artifactId>
     <version>2.1.1-RC-SNAPSHOT</version>
     <relativePath>../headless-server-commerce-parent</relativePath>
   </parent>
-
   <artifactId>headless-server-commerce-lib</artifactId>
-  <description>
-    The implementation of the Commerce Headless Server.
-    Can be started as a separate application or integrated as a CMS project extension.
-  </description>
+  <description>The implementation of the Commerce Headless Server.
+    Can be started as a separate application or integrated as a CMS project extension.</description>
 
   <properties>
     <!--
@@ -79,9 +75,12 @@
       <artifactId>spring-context</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-core</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-autoconfigure</artifactId>
     </dependency>
   </dependencies>
-
 </project>
