+++ headless-server-commerce-parent/pom.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
   <parent>
@@ -37,14 +36,15 @@
     <maven.test.redirectTestOutputToFile>true</maven.test.redirectTestOutputToFile>
 
     <!-- versions managed in project -->
-    <cms.version>2401.3</cms.version>
-    <graphql-spring-boot-starter.version>15.1.0</graphql-spring-boot-starter.version>
+    <cms.version>2404.1</cms.version>
     <jetty.version>12.0.6</jetty.version>
     <pf4j.version>3.11.0</pf4j.version>
     <sonar-maven-plugin.version>3.11.0.3922</sonar-maven-plugin.version>
     <sortpom-maven-plugin.version>3.4.1</sortpom-maven-plugin.version>
     <springfox.version>3.0.0</springfox.version>
     <coremedia-enforcer-rules.version>1.2.2</coremedia-enforcer-rules.version>
+    <log4j2slf4j.version>2.21.1</log4j2slf4j.version>
+    <slf4j.version>2.0.12</slf4j.version>
 
     <!-- versions managed in spring-boot-starter-parent (overwrites) -->
     <graphql-java.version>21.3</graphql-java.version>
@@ -170,37 +170,10 @@
         <artifactId>graphql-java</artifactId>
         <version>${graphql-java.version}</version>
       </dependency>
-      <!-- to embed GraphiQL tool -->
       <dependency>
-        <groupId>com.graphql-java-kickstart</groupId>
-        <artifactId>graphql-spring-boot-starter</artifactId>
-        <version>${graphql-spring-boot-starter.version}</version>
-        <exclusions>
-          <exclusion>
-            <groupId>com.google.code.findbugs</groupId>
-            <artifactId>jsr305</artifactId>
-          </exclusion>
-          <exclusion>
-            <groupId>com.graphql-java-kickstart</groupId>
-            <artifactId>graphql-java-servlet</artifactId>
-          </exclusion>
-          <exclusion>
-            <groupId>com.graphql-java-kickstart</groupId>
-            <artifactId>graphql-java-tools</artifactId>
-          </exclusion>
-          <exclusion>
-            <groupId>io.github.graphql-java</groupId>
-            <artifactId>graphql-java-annotations</artifactId>
-          </exclusion>
-          <exclusion>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
-          </exclusion>
-          <exclusion>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-validation</artifactId>
-          </exclusion>
-        </exclusions>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-web</artifactId>
+        <scope>runtime</scope>
       </dependency>
       <dependency>
         <groupId>io.springfox</groupId>
@@ -243,6 +216,23 @@
         <groupId>org.pf4j</groupId>
         <artifactId>pf4j</artifactId>
         <version>${pf4j.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-bom</artifactId>
+        <version>${slf4j.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-to-slf4j</artifactId>
+        <version>${log4j2slf4j.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.logging.log4j</groupId>
+        <artifactId>log4j-api</artifactId>
+        <version>${log4j2slf4j.version}</version>
       </dependency>
       <dependency>
         <groupId>org.yaml</groupId>
@@ -314,8 +304,13 @@
         <plugin>
           <groupId>com.google.cloud.tools</groupId>
           <artifactId>jib-maven-plugin</artifactId>
-          <version>3.4.2</version>
+          <version>3.4.0</version>
           <dependencies>
+            <dependency>
+              <groupId>com.google.cloud.tools</groupId>
+              <artifactId>jib-layer-filter-extension-maven</artifactId>
+              <version>0.3.0</version>
+            </dependency>
             <dependency>
               <groupId>com.google.cloud.tools</groupId>
               <artifactId>jib-ownership-extension-maven</artifactId>
